Separate classes for Ui, TaskList, Storage
