+++ src/main/java/Ui.java
@@ -1,0 +1,11 @@
+class Ui {
+    private static final String INDENTATION = "    ";
+
+    public void outputText(String text) {
+        System.out.println(INDENTATION + " " + text);
+    }
+
+    public void printSeparatingLine() {
+        System.out.println(INDENTATION + "____________________________________________________________");
+    }
+}
